Changed UI system to use floats
Prevent conversions when interacting with Graphics and Drawable.

+++ CompGeoAlgoVisualizer/Source/UI/IButton.cpp
@@ -5,13 +5,13 @@
 
 using namespace D2D;
 
-IButton::IButton(std::wstring text, int width, int height)
+IButton::IButton(std::wstring text, float width, float height)
 	:
 	IButton(text, TextFormat(L"Arial", 14, true, Style::NORMAL, TextAlignment::CENTER, ParagraphAlignment::CENTER), width, height)
 {
 }
 
-IButton::IButton(std::wstring text, TextFormat& textFormat, int width, int height)
+IButton::IButton(std::wstring text, TextFormat& textFormat, float width, float height)
 	:
 	IComponent(0, 0, width, height),
 	pText(new Text(text, textFormat, width, height)),
@@ -36,21 +36,21 @@ void IButton::setTextColor(const Color& color)
 	pText->setColor(color);
 }
 
-void IButton::setWidth(int width)
+void IButton::setWidth(float width)
 {
 	IComponent::setWidth(width);
 	pRectangle->setWidth(width);
 	pText->setWidth(width);
 }
 
-void IButton::setHeight(int height)
+void IButton::setHeight(float height)
 {
 	IComponent::setHeight(height);
 	pRectangle->setHeight(height);
 	pText->setHeight(height);
 }
 
-void IButton::onHover(int x, int y, MouseEvent& mouseEvent)
+void IButton::onHover(float x, float y, MouseEvent& mouseEvent)
 {
 	if (mouseEvent.isConsumed)
 		return;
@@ -68,14 +68,14 @@ void IButton::onExit()
 	pText->setColor(mTextColor);
 }
 
-void IButton::onPress(int x, int y, MouseEvent& mouseEvent)
+void IButton::onPress(float x, float y, MouseEvent& mouseEvent)
 {
 	if (mouseEvent.isConsumed)
 		return;
 	mouseEvent.isConsumed = true;
 }
 
-void IButton::onClick(int x, int y, MouseEvent& mouseEvent)
+void IButton::onClick(float x, float y, MouseEvent& mouseEvent)
 {
 	if (mouseEvent.isConsumed)
 		return;
@@ -88,12 +88,12 @@ void IButton::setOnClick(const std::function<void()>& handler)
 	mOnClickHandler = handler;
 }
 
-void IButton::setCornerRadius(int radius)
+void IButton::setCornerRadius(float radius)
 {
 	pRectangle->setCornerRadius(radius);
 }
 
-bool IButton::inComponent(int x, int y)
+bool IButton::inComponent(float x, float y)
 {
-	return pRectangle->inRect({ (float)x, (float)y });
+	return pRectangle->inRect({ x, y });
 }
